Refaktorer og flytt hooks inn i komponent
